finished routes and basic controllers

+++ backend/controllers/triggerController.js
@@ -13,12 +13,23 @@ module.exports.createTrigger = asyncHandler(async (req, res) => {
 });
 
 module.exports.updateTrigger = asyncHandler(async (req, res) => {
-    const trigger = await Trigger.findByIdAndUpdate({ _id: req.params.id }, req.body);
+    const trigger = Trigger.findById({ _id: req.params.id });
+
+    const user = await User.findById(req.user.id);
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
     if (!trigger) {
         res.status(400);
         throw new Error("trigger not found");
     }
-    res.status(200).json(trigger);
+    if (trigger.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
+    const updatedTrigger = await Trigger.findByIdAndUpdate({ _id: req.params.id }, req.body);
+    res.status(200).json(updatedTrigger);
 });
 
 module.exports.singleTrigger = asyncHandler(async (req, res) => {
@@ -31,11 +42,25 @@ module.exports.singleTrigger = asyncHandler(async (req, res) => {
 });
 
 module.exports.deleteTrigger = asyncHandler(async (req, res) => {
-    const trigger = await Trigger.findByIdAndRemove({ _id: req.params.id });
+    const trigger = await Trigger.findById({ _id: req.params.id });
+    const user = await User.findById(req.user.id);
 
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+    if (!trigger) {
+        res.status(400);
+        throw new Error("trigger not found");
+    }
+    if (trigger.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
     if (!trigger) {
         res.status(400);
         throw new Error("Cannot find trigger");
     }
-    res.status(200).json(trigger);
+    const deletedTrigger = await Trigger.findByIdAndRemove({ _id: req.params.id });
+    res.status(200).json(deletedTrigger);
 });
